Menambahkan syarat command cc agar terlarang bagi role selain admin

+++ commands/clear-chat.js
@@ -2,11 +2,17 @@ module.exports = {
     name: 'clear-chat',
     description: 'Command untuk !clear-chat',
     execute(message, args) {
-        if (!args[1]) {
-            message.reply('masukkan jumlah chat yang akan dihapus setelah !clear-chat');
+        const role = message.member.roles.cache.find(r => r.name === 'Admin');
+        if (role) {
+            if (!args[1]) {
+                message.reply('masukkan jumlah chat yang akan dihapus setelah !clear-chat');
+            }
+            else {
+                message.channel.bulkDelete(args[1]);
+            }
         }
         else {
-            message.channel.bulkDelete(args[1]);
+            message.reply('siapa kamu wahai kisanak, kamu terlarang menggunakan command ini');
         }
     },
 };
